+++ hdf5io.h
@@ -8,15 +8,15 @@
 struct hdf5io_waveform_file 
 {
     hid_t waveFid;
-    size_t nPt;
-    size_t nCh;
-    size_t nWfmPerChunk;
-    size_t nEvents;
+    uint64_t nPt;
+    uint64_t nCh;
+    uint64_t nWfmPerChunk;
+    uint64_t nEvents;
 };
 
 struct hdf5io_waveform_event
 {
-    size_t eventId;
+    uint64_t eventId;
     /* wavBuf should point to a contiguous 2D array, mapped as
      * ch1..ch2..ch3..ch4 (row-major).  Omitting one or more ch? is
      * allowed in accordance with chMask.*/
@@ -27,8 +27,8 @@ struct hdf5io_waveform_event
  * performance, n waveforms are grouped together to be put in the same
  * array, then the (n+1)th waveform is put into the next grouped
  * array, and so forth. */
-struct hdf5io_waveform_file *hdf5io_open_file(const char *fname, size_t nWfmPerChunk,
-                                              size_t nCh);
+struct hdf5io_waveform_file *hdf5io_open_file(const char *fname, uint64_t nWfmPerChunk,
+                                              uint64_t nCh);
 struct hdf5io_waveform_file *hdf5io_open_file_for_read(const char *fname);
 int hdf5io_close_file(struct hdf5io_waveform_file *wavFile);
 /* flush also writes nEvents to the file */
@@ -42,6 +42,7 @@ int hdf5io_write_event(struct hdf5io_waveform_file *wavFile,
                        struct hdf5io_waveform_event *wavEvent);
 int hdf5io_read_event(struct hdf5io_waveform_file *wavFile,
                       struct hdf5io_waveform_event *wavEvent);
-size_t hdf5io_get_number_of_events(struct hdf5io_waveform_file *wavFile);
+uint64_t hdf5io_get_number_of_events(struct hdf5io_waveform_file *wavFile);
 
 #endif /* __HDF5IO_H__ */
+
